src/test/kotlin/net/tlalka/puzzles/aoc2021/Day04Test.kt
package net.tlalka.puzzles.aoc2021

import net.tlalka.puzzles.common.extension.puzzleFile
import net.tlalka.puzzles.common.extension.readInput
import net.tlalka.puzzles.common.extension.sampleFile
import org.junit.jupiter.api.MethodOrderer.MethodName
import org.junit.jupiter.api.Test
import org.junit.jupiter.api.TestMethodOrder
import kotlin.test.assertEquals

@TestMethodOrder(MethodName::class)
class Day04Test {

    private val tested = Day04()

    private val sampleInput = sampleFile<Day04>().readInput()
    private val puzzleInput = puzzleFile<Day04>().readInput()

    @Test
    fun `verify part 1 with sample input`() {
        assertEquals(4512, tested.part1(sampleInput))
    }

    @Test
    fun `verify part 1 with puzzle input`() {
        assertEquals(8442, tested.part1(puzzleInput))
    }

    @Test
    fun `verify part 2 with sample input`() {
        assertEquals(1924, tested.part2(sampleInput))
    }

    @Test
    fun `verify part 2 with puzzle input`() {
        assertEquals(4590, tested.part2(puzzleInput))
    }
}
